requirements.txt
pandasai
pandasai[excel]
pandasai[connectors]
tiktoken
streamlit
numpy==1.23.5 
pandas==1.5.3
pyyaml
seaborn
google-generativeai